+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add new requirements battery not low and storage not low
 * Add helper job class `DailyJob` to make it easier to run jobs once a day, see #223
 * Add option in `JobConfig` to add a logger
+* Add option in `JobConfig` for a job ID offset to avoid clashes with other jobs in the `JobScheduler`
 
 ## 1.1.12